requirements.txt
altgraph==0.17
auto-py-to-exe==2.7.4
bottle==0.12.18
bottle-websocket==0.2.9
cairocffi==1.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
decorator==4.4.2
Eel==0.11.0
future==0.18.2
gevent==20.6.0
gevent-websocket==0.10.1
gizeh==0.1.11
greenlet==0.4.16
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.2
moviepy==1.0.3
numpy==1.18.5
opencv-python==4.2.0.34
pefile==2019.4.18
Pillow==7.1.2
proglog==0.1.9
pycparser==2.20
PyInstaller==3.6
pywin32-ctypes==0.2.0
requests==2.23.0
tqdm==4.46.1
urllib3==1.25.9
whichcraft==0.6.1
zope.event==4.4
zope.interface==5.1.0
